Rename core-component to core and move inside components
