+++ src/components/core/sidebar/sidebar.tsx
@@ -19,7 +19,7 @@
 import React, { FunctionComponent, ReactElement } from "react";
 import { AiOutlineCopyrightCircle } from "react-icons/all";
 import { Template } from "./sidebar-template";
-import templates from "../../api/templates.json";
+import templates from "../../../api/templates.json";
 
 /**
  * Sidebar component.
